backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
websockets==10.4
pydantic==1.10.7
python-dateutil==2.8.2
